+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased:
   * Add `Buffer::new_owned_preallocated(...)` to transfer ownership of a pre-allocated buffer to Aravis.
+  * Deprecate `Buffer::new_preallocated(...)` in favor of `Buffer::new_preallocated_borrowed`.
 
 v0.8.0:
   * Update to `gtk-rs` v0.15.